requirements.txt
asgiref==3.8.1
Django==5.1.6
django-allauth==65.4.1
django-ckeditor==6.7.2
django-js-asset==3.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
